+++ naga/src/front/wgsl/error.rs
@@ -263,7 +263,7 @@ pub(crate) enum Error<'a> {
     },
     FunctionReturnsVoid(Span),
     FunctionMustUseUnused(Span),
-    FunctionMustUseNothing(Span, Span),
+    FunctionMustUseReturnsVoid(Span, Span),
     InvalidWorkGroupUniformLoad(Span),
     Internal(&'static str),
     ExpectedConstExprConcreteIntegerScalar(Span),
@@ -823,27 +823,23 @@ impl<'a> Error<'a> {
                 ],
             },
             Error::FunctionMustUseUnused(call) => ParseError {
-                message: "must use result".into(),
-                labels: vec![(call, "result unused".into())],
+                message: "unused return value from function annotated with @must_use".into(),
+                labels: vec![(call, "".into())],
                 notes: vec![
                     format!(
                         "function '{}' is declared with `@must_use` attribute",
-                        &source[call].split_once('(').unwrap().0
+                        &source[call],
                     ),
-                    "call the function as part of an expression".into(),
+                    "use a phony assignment or declare a value using the function call as the initializer".into(),
                 ],
             },
-            Error::FunctionMustUseNothing(attr, signature) => ParseError {
-                message: "must use nothing".into(),
+            Error::FunctionMustUseReturnsVoid(attr, signature) => ParseError {
+                message: "function annotated with @must_use but does not return any value".into(),
                 labels: vec![
-                    (attr, "requires result".into()),
-                    (signature, "no result".into()),
+                    (attr, "".into()),
+                    (signature, "".into()),
                 ],
                 notes: vec![
-                    format!(
-                        "function '{}' is declared with `@must_use` attribute",
-                        &source[signature].split_once('(').unwrap().0
-                    ),
                     "declare a return type or remove the attribute".into(),
                 ],
             },