+++ naga/src/valid/function.rs
@@ -22,8 +22,6 @@ pub enum CallError {
     ResultAlreadyInScope(Handle<crate::Expression>),
     #[error("Result expression {0:?} is populated by multiple `Call` statements")]
     ResultAlreadyPopulated(Handle<crate::Expression>),
-    #[error("Result expression {0:?} is not used")]
-    ResultNotUsed(Handle<crate::Expression>),
     #[error("Result value is invalid")]
     ResultValue(#[source] ExpressionError),
     #[error("Requires {required} arguments, but {seen} are provided")]
@@ -224,7 +222,6 @@ struct BlockContext<'a> {
     abilities: ControlFlowAbility,
     info: &'a FunctionInfo,
     expressions: &'a Arena<crate::Expression>,
-    named_expressions: &'a crate::NamedExpressions,
     types: &'a UniqueArena<crate::Type>,
     local_vars: &'a Arena<crate::LocalVariable>,
     global_vars: &'a Arena<crate::GlobalVariable>,
@@ -245,7 +242,6 @@ impl<'a> BlockContext<'a> {
             abilities: ControlFlowAbility::RETURN,
             info,
             expressions: &fun.expressions,
-            named_expressions: &fun.named_expressions,
             types: &module.types,
             local_vars: &fun.local_variables,
             global_vars: &module.global_variables,
@@ -335,14 +331,6 @@ impl super::Validator {
                 crate::Expression::CallResult(callee)
                     if fun.result.is_some() && callee == function =>
                 {
-                    if fun.result.as_ref().unwrap().must_use
-                        && context.info[expr].ref_count == 0
-                        && context.named_expressions.get(&expr).is_none()
-                    {
-                        return Err(CallError::ResultNotUsed(expr)
-                            .with_span_handle(expr, context.expressions));
-                    }
-
                     if !self.needs_visit.remove(expr) {
                         return Err(CallError::ResultAlreadyPopulated(expr)
                             .with_span_handle(expr, context.expressions));
@@ -1580,6 +1568,7 @@ impl super::Validator {
         fun_info: &FunctionInfo,
         local_expr_kind: &crate::proc::ExpressionKindTracker,
     ) -> Result<(), LocalVariableError> {
+        log::debug!("var {:?}", var);
         let type_info = self
             .types
             .get(var.ty.index())